Minor improvemnts for app page
Most likely those are the last changes made for the app page. They are
the cherry on top of recent work around this View. Here are the
improvemnts:

1) Each weather box is now non-draggable. That means user cannot
move a linked box to a new tab. The only way is to click onto a
weather box (which prevents from crashing the app).

2) Changed hover background for "cloudy" type of weather, so that it
has more balanced, stable background.

3) Animated hover effect so that the background appears when user
keeps the mouse cursor onto a weather box. Also once cursor leaves
the weather box, fading animation is fired, causing the background
to disappear smoothly.