Include footer to a webpage
This update adds a brand new footer to the very bottom of detailed
webpage. It is divided into 3 sections. First one is 'About the project',
which verbosely describes the main prupose of the app and includes the
copyright note. Second one is handling some extra important reference
links related to the app (documentation, changelog, license and credits
sections) - they currently do not redirect anywhere. Lastly, a 'follow'
section, containing 4 social media boxes, which will redirect to a given
social media account of the app author.