Finish queries for detailed weather page
It's finally done ! All media queries are fully prepared for the website.
Here a tablet and portrait oriented mobiles got served with a lot of
corrections, which as a result fit them appropiately to the content
that websites offer. Minor corrections are still to do (they are some
further improvements expected), however for now the app content seems
to be scaled properly.